Refactor: tabs text
